refactor(ts): give FormatPipeline its final DI shape and a fluent pipeline
Move format-pipeline.ts into pipeline/, extract validate() into a SyntaxValidator
(deps { sourceFiles, splitter, parser }), and reduce FormatPipeline to
{ editor, processRunner, validator } with a runPass(FileFormatter, files, mode)
generic over both pipelines. Add PipelineFactory.fluentPipeline() composing
fluent-chain, Drizzle-query, and expanded-call passes in the exact order the old
FluentChains.format used, plus segmentFormatter/fluentFormatter/syntaxValidator
builders so the composition root vends the whole formatting graph.
