--- packages/ts/sidecar/src/format-pipeline.ts
@@ -1,222 +1,0 @@
-import { availableParallelism } from 'node:os';
-import { EmbeddedBlockSplitter } from '#sidecar/hosts/embedded-block-splitter';
-import type { OxfmtRunFailed, SourceFileUnreadable, SourceUnparsable } from '#sidecar/kernel/errors';
-import { FileFormatter } from '#sidecar/pipeline/file-formatter';
-import { FluentChains } from '#sidecar/fluent-chains';
-import { PipelineFactory } from '#sidecar/pipeline/pipeline-factory';
-import type { ProcessRunner } from '#sidecar/io/process-runner';
-import { isErr, ok } from '#sidecar/kernel/result';
-import type { Result } from '#sidecar/kernel/result';
-import { SourceFileEditor } from '#sidecar/pipeline/source-file-editor';
-import type { SourceFileError, SourceFiles } from '#sidecar/io/source-files';
-import { SourceParser } from '#sidecar/syntax/source-parser';
-
-const OXFMT_CHUNK_SIZE = 100;
-
-/** Whether a pipeline pass checks source or writes its computed changes. */
-export type FormatMode = 'check' | 'write';
-
-/** The result of processing one file in a formatting pass. */
-export type PassOutcome = {
-	/** The formatting pass that produced the outcome. */
-	readonly label: string;
-
-	/** The requested source path. */
-	readonly file: string;
-
-	/** Whether the pass would change or did change the source. */
-	readonly changed: boolean;
-
-	/** The typed filesystem failure, or `null` when processing completed. */
-	readonly error: SourceFileError | null;
-};
-
-/** The options needed to invoke oxfmt over one pipeline stage. */
-export type OxfmtOptions = {
-	/** The executable to invoke, or `null` to skip the stage. */
-	readonly bin: string | null;
-
-	/** The oxfmt configuration path, or `null` to use defaults. */
-	readonly config: string | null;
-
-	/** The source paths passed to oxfmt. */
-	readonly files: string[];
-
-	/** Whether oxfmt checks source or writes changes. */
-	readonly mode: FormatMode;
-};
-
-/** A source file that could not be read or parsed during validation. */
-export type ValidationFailure = {
-	/** The original source path reported to the user. */
-	readonly file: string;
-
-	/** The carried read or parse failure. */
-	readonly error: SourceFileUnreadable | SourceUnparsable;
-};
-
-type ProcessOne = (file: string, mode: FormatMode) => Promise<Result<boolean, SourceFileError>>;
-
-/** Coordinates formatting and validation through narrow filesystem and process ports. */
-export class FormatPipeline {
-	readonly #sourceFiles: SourceFiles;
-	readonly #processRunner: ProcessRunner;
-	readonly #parser: SourceParser;
-	readonly #splitter: EmbeddedBlockSplitter;
-	readonly #editor: SourceFileEditor;
-	readonly #fileFormatter: FileFormatter;
-
-	static async #mapPool<T, R>(items: T[], limit: number, operation: (item: T) => Promise<R>): Promise<R[]> {
-		const results = new Array<R>(items.length);
-
-		let nextIndex = 0;
-
-		const worker = async (): Promise<void> => {
-			while (true) {
-				const index = nextIndex++;
-
-				if (index >= items.length) {
-					return;
-				}
-
-				const item = items[index];
-
-				if (item !== undefined) {
-					results[index] = await operation(item);
-				}
-			}
-		};
-
-		const workers = Array.from({ length: Math.max(1, Math.min(limit, items.length)) }, worker);
-
-		await Promise.all(workers);
-
-		return results;
-	}
-
-	static #scriptPrefix(content: string, scriptStart: number): string {
-		return content.slice(0, scriptStart).replace(/[^\r\n]/g, ' ');
-	}
-
-	/**
-	 * @param dependencies - The filesystem and process ports used by the pipeline.
-	 * @param dependencies.sourceFiles - Reads and atomically writes source files.
-	 * @param dependencies.processRunner - Invokes oxfmt with inherited standard streams.
-	 */
-	constructor(dependencies: { sourceFiles: SourceFiles; processRunner: ProcessRunner }) {
-		this.#sourceFiles = dependencies.sourceFiles;
-		this.#processRunner = dependencies.processRunner;
-		this.#parser = new SourceParser();
-		this.#splitter = new EmbeddedBlockSplitter();
-		this.#editor = new SourceFileEditor({ sourceFiles: dependencies.sourceFiles });
-		this.#fileFormatter = new FileFormatter({ splitter: this.#splitter, pipeline: PipelineFactory.create().segmentPipeline() });
-	}
-
-	/**
-	 * Apply the blank-line formatting pass to one TypeScript or host file.
-	 *
-	 * @param path - The source file to format.
-	 * @param mode - Whether to check or atomically write changes.
-	 * @returns Whether the file changes, or the typed filesystem failure.
-	 */
-	async formatFile(path: string, mode: FormatMode): Promise<Result<boolean, SourceFileError>> {
-		return this.#editor.apply(path, mode, (content) => {
-			return this.#fileFormatter.format(path, content);
-		});
-	}
-
-	/**
-	 * Apply fluent-chain formatting to one TypeScript or host file.
-	 *
-	 * @param path - The source file to format.
-	 * @param mode - Whether to check or atomically write changes.
-	 * @returns Whether the file changes, or the typed filesystem failure.
-	 */
-	formatFluentFile(path: string, mode: FormatMode): Promise<Result<boolean, SourceFileError>> {
-		return FluentChains.formatFile(path, mode, this.#sourceFiles);
-	}
-
-	/**
-	 * Process files concurrently while preserving outcome order.
-	 *
-	 * @param label - The pass label associated with the outcomes.
-	 * @param files - The source paths to process.
-	 * @param mode - Whether the pass checks or writes changes.
-	 * @param processOne - The operation applied to each source path.
-	 * @returns One effect-free reporting outcome per input path.
-	 */
-	async runPass(label: string, files: string[], mode: FormatMode, processOne: ProcessOne): Promise<PassOutcome[]> {
-		return FormatPipeline.#mapPool(files, availableParallelism(), async (file): Promise<PassOutcome> => {
-			const outcome = await processOne(file, mode);
-
-			if (isErr(outcome)) {
-				return { label, file, changed: false, error: outcome.error };
-			}
-
-			return { label, file, changed: outcome.value, error: null };
-		});
-	}
-
-	/**
-	 * Run oxfmt sequentially over bounded file chunks.
-	 *
-	 * @param options - The executable, configuration, files, and format mode.
-	 * @returns Nothing, or the first typed oxfmt failure.
-	 */
-	async runOxfmt(options: OxfmtOptions): Promise<Result<void, OxfmtRunFailed>> {
-		if (!options.bin || options.files.length === 0) {
-			return ok(undefined);
-		}
-
-		const args = options.config ? ['--config', options.config] : [];
-
-		args.push(options.mode === 'check' ? '--check' : '--write', '--no-error-on-unmatched-pattern');
-
-		for (let i = 0; i < options.files.length; i += OXFMT_CHUNK_SIZE) {
-			const outcome = await this.#processRunner.run(options.bin, [...args, ...options.files.slice(i, i + OXFMT_CHUNK_SIZE)]);
-
-			if (isErr(outcome)) {
-				return outcome;
-			}
-		}
-
-		return ok(undefined);
-	}
-
-	/**
-	 * Validate TypeScript files and JavaScript-compatible embedded host blocks.
-	 *
-	 * @param files - The source paths to validate.
-	 * @returns Carried read and parse failures in deterministic input order.
-	 */
-	async validate(files: string[]): Promise<ValidationFailure[]> {
-		const failures = await FormatPipeline.#mapPool(files, availableParallelism(), async (file): Promise<ValidationFailure[]> => {
-			const read = await this.#sourceFiles.readText(file);
-
-			if (isErr(read)) {
-				return [{ file, error: read.error }];
-			}
-
-			if (!this.#splitter.isHost(file)) {
-				const parsed = this.#parser.parse(file, read.value);
-
-				return isErr(parsed) ? [{ file, error: parsed.error }] : [];
-			}
-
-			const hostFailures: ValidationFailure[] = [];
-
-			for (const block of this.#splitter.extract(file, read.value)) {
-				const virtualContent = FormatPipeline.#scriptPrefix(read.value, block.start) + block.content;
-				const parsed = this.#parser.parse(`${file}.script.${block.extension}`, virtualContent);
-
-				if (isErr(parsed) && this.#splitter.hardValidated(file)) {
-					hostFailures.push({ file, error: parsed.error });
-				}
-			}
-
-			return hostFailures;
-		});
-
-		return failures.flat();
-	}
-}
